AsyncScrolling: add ScrollingTreePositionedNode base class https://bugs.webkit.org/show_bug.cgi?id=246453
Reviewed by Žan Doberšek.

Make Cocoa and Nicosia implementations inherit from the base class and move the common code there.

* Source/WebCore/Sources.txt:
* Source/WebCore/SourcesCocoa.txt:
* Source/WebCore/SourcesGTK.txt:
* Source/WebCore/SourcesWPE.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/page/scrolling/ScrollingTreePositionedNode.cpp: Renamed from Source/WebCore/page/scrolling/cocoa/ScrollingTreePositionedNode.mm.
(WebCore::ScrollingTreePositionedNode::commitStateBeforeChildren):
(WebCore::ScrollingTreePositionedNode::scrollDeltaSinceLastCommit const):
* Source/WebCore/page/scrolling/ScrollingTreePositionedNode.h: Copied from Source/WebCore/page/scrolling/cocoa/ScrollingTreePositionedNode.h.
* Source/WebCore/page/scrolling/cocoa/ScrollingTreeFixedNode.mm:
(WebCore::ScrollingTreeFixedNode::applyLayerPositions):
* Source/WebCore/page/scrolling/cocoa/ScrollingTreePositionedNodeCocoa.h: Renamed from Source/WebCore/page/scrolling/cocoa/ScrollingTreePositionedNode.h.
(WebCore::ScrollingTreePositionedNodeCocoa::layer const):
* Source/WebCore/page/scrolling/cocoa/ScrollingTreePositionedNodeCocoa.mm: Added.
(WebCore::ScrollingTreePositionedNodeCocoa::create):
(WebCore::ScrollingTreePositionedNodeCocoa::ScrollingTreePositionedNodeCocoa):
(WebCore::ScrollingTreePositionedNodeCocoa::commitStateBeforeChildren):
(WebCore::ScrollingTreePositionedNodeCocoa::applyLayerPositions):
* Source/WebCore/page/scrolling/mac/ScrollingTreeMac.mm:
(ScrollingTreeMac::createScrollingTreeNode):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeFixedNode.cpp:
(WebCore::ScrollingTreeFixedNode::applyLayerPositions):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeNicosia.cpp:
(WebCore::ScrollingTreeNicosia::createScrollingTreeNode):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreePositionedNodeNicosia.cpp: Renamed from Source/WebCore/page/scrolling/nicosia/ScrollingTreePositionedNode.cpp.
(WebCore::ScrollingTreePositionedNodeNicosia::create):
(WebCore::ScrollingTreePositionedNodeNicosia::ScrollingTreePositionedNodeNicosia):
(WebCore::ScrollingTreePositionedNodeNicosia::commitStateBeforeChildren):
(WebCore::ScrollingTreePositionedNodeNicosia::applyLayerPositions):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreePositionedNodeNicosia.h: Renamed from Source/WebCore/page/scrolling/nicosia/ScrollingTreePositionedNode.h.
* Source/WebKit/UIProcess/RemoteLayerTree/RemoteScrollingTree.cpp:
(WebKit::RemoteScrollingTree::createScrollingTreeNode):

Canonical link: https://commits.webkit.org/255724@main
